Refactors the redcarpet_code_renderer mixin
I was getting strange errors with middleman when a page was html.md.erb
versus html.md. I tracked the issue down to there being a problem with
the last match group not properly being filled so I changed the match
criteria and take in the whole code block.

Instead of polluting the Redcarpet I moved that all to a new module so
the incision point is smaller.

The TerminalParser is in a temporary state. It unlexes the code and essentially
lexes each line instead. That way the output is not lexed and there is more
control on continuations.

Signed-off-by: Franklin Webber <franklin@chef.io>