Simplifies Highlighter
The highlighter is no longer a weird switch board generating lots
of instances or having a huge table. Changed to either of the two
formatters now and that is it.

Moved the lexing of the code to the code formatter
Removed the unlexing of the code in the terminal formatter
Added a habtiat studio lexer that adds itself to rouge

Signed-off-by: Franklin Webber <franklin@chef.io>